Fix rest/transform package moving
